added repo, share and solr warpath, in order to deploy war files from local FS into the vms, instead of using maven to remotely fetch them

+++ recipes/share.rb
@@ -18,6 +18,7 @@
 #
 maven_repos               = node['alfresco']['maven']['repos']
 
+share_warpath             = node['alfresco']['share']['war_path']
 share_groupId             = node['alfresco']['share']['groupId']
 share_artifactId          = node['alfresco']['share']['artifactId']
 share_version             = node['alfresco']['share']['version']
@@ -64,31 +65,42 @@ maven "share" do
   subscribes    :put, "template[share-config-custom.xml]", :immediately
 end
 
-ark "share" do
-  url                 "file://#{cache_path}/share.war"
-  path                cache_path
-  owner               alfresco_user
-  action              :put
-  strip_leading_dir   false
-  append_env_path     false
-  subscribes          :put, "maven[share]", :immediately
-end
-
-template "share-log4j" do
-  path        "#{cache_path}/share/WEB-INF/classes/log4j.properties"
-  source      "share-log4j.properties.erb"
-  owner       alfresco_user
-  group       alfresco_group
-  mode        "0664"
-  subscribes  :create, "ark[share]", :immediately
-end
-
-ruby_block "deploy-share" do
-  block do
-    require 'fileutils'
-    FileUtils.rm_rf "#{cache_path}/share/share"
-    FileUtils.cp_r "#{cache_path}/share","#{webapp_dir}/share"
+if !share_warpath.nil?
+  # Deploy the war file as it is
+  ruby_block "deploy-share-warpath" do
+    block do
+      require 'fileutils'
+      FileUtils.cp "#{share_warpath}","#{webapp_dir}/"
+    end
+    notifies :restart, "service[tomcat]"
+  end  
+else
+  ark "share" do
+    url                 "file://#{cache_path}/share.war"
+    path                cache_path
+    owner               alfresco_user
+    action              :put
+    strip_leading_dir   false
+    append_env_path     false
+    subscribes          :put, "maven[share]", :immediately
   end
-  subscribes  :create, "template[share-log4j]", :immediately
-  notifies    :restart, "service[tomcat]"
+
+  template "share-log4j" do
+    path        "#{cache_path}/share/WEB-INF/classes/log4j.properties"
+    source      "share-log4j.properties.erb"
+    owner       alfresco_user
+    group       alfresco_group
+    mode        "0664"
+    subscribes  :create, "ark[share]", :immediately
+  end
+
+  ruby_block "deploy-share" do
+    block do
+      require 'fileutils'
+      FileUtils.rm_rf "#{cache_path}/share/share"
+      FileUtils.cp_r "#{cache_path}/share","#{webapp_dir}/share"
+    end
+    subscribes  :create, "template[share-log4j]", :immediately
+    notifies    :restart, "service[tomcat]"
+  end
 end
